@@ -1,13 +1,6 @@
-module.exports = {
-  plugins: [
-    require('postcss-flexbugs-fixes'),
-    require('postcss-preset-env')({
-      autoprefixer: {
-        flexbox: 'no-2009',
-      },
-      stage: 3,
-    }),
-    require('tailwindcss'),
-    require('autoprefixer')
-  ],
-};
+export default {
+  plugins: {
+    '@tailwindcss/postcss': {},
+    autoprefixer: {},
+  },
+}
