@@ -1,18 +1,10 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: [
-        "./src/**/*.{js,jsx,ts,tsx,html}",
-    ],
-    theme: {
-        extend: {
-            colors: {
-                mikrotik: {
-                    blue: '#1e40af',
-                    green: '#059669',
-                    red: '#dc2626'
-                }
-            }
-        },
-    },
-    plugins: [],
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
 }
